Moved adding userId to collaborator to an extension method
